reorganize dag.py
